+++ src/megawrapper/__init__.py
@@ -1,6 +1,7 @@
 from .board import Board
 from .motor import Motor
 from .servo import Servo
+from .tcs34725 import TCS34725
 from .utils import delay, millis
 
 from .exceptions import (
@@ -8,6 +9,7 @@ from .exceptions import (
     InvalidSpeedError,
     BoardConnectionError,
     StandbyNotConfiguredError,
+    TCS34725Error,
 )
 
 from .version import __version__
@@ -16,11 +18,13 @@ __all__ = [
     "Board",
     "Motor",
     "Servo",
+    "TCS34725",
     "delay",
     "millis",
     "MegaWrapperError",
     "InvalidSpeedError",
     "BoardConnectionError",
     "StandbyNotConfiguredError",
+    "TCS34725Error",
     "__version__",
 ]